+++ example/sse-server.js
@@ -1,27 +1,29 @@
-var express = require('express')
-var serveStatic = require('serve-static')
-var SSE = require('sse')
+const express = require('express')
+const serveStatic = require('serve-static')
+const SseStream = require('ssestream')
 
-var app = express()
+const app = express()
 app.use(serveStatic(__dirname))
-
-var server = app.listen(8080, function (err) {
-  if (err) throw err
-  console.log('server ready on http://localhost:8080')
-})
-
-var sse = new SSE(server)
-sse.on('connection', function (connection) {
+app.get('/sse', (req, res) => {
   console.log('new connection')
-  var pusher = setInterval(function () {
-    connection.send({
+
+  const sseStream = new SseStream(req)
+  sseStream.pipe(res)
+  const pusher = setInterval(() => {
+    sseStream.write({
       event: 'server-time',
       data: new Date().toTimeString()
     })
   }, 1000)
 
-  connection.on('close', function () {
+  res.on('close', () => {
     console.log('lost connection')
     clearInterval(pusher)
+    sseStream.unpipe(res)
   })
+})
+
+app.listen(8080, (err) => {
+  if (err) throw err
+  console.log('server ready on http://localhost:8080')
 })